// src/domain/enums/saga-step.enum.ts
export enum StepStatus {
  PROCESSING = 'PROCESSING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  COMPENSATING = 'COMPENSATING',
  COMPENSATED = 'COMPENSATED',
}

export enum SagaStepName {
  VALIDATE_ITEMS = 'VALIDATE_ITEMS',
  GET_ADDRESS = 'GET_ADDRESS',
  VALIDATE_VOUCHERS = 'VALIDATE_VOUCHERS',
  CALCULATE_AND_VERIFY_PRICE = 'CALCULATE_AND_VERIFY_PRICE',
  RESERVE_INVENTORY = 'RESERVE_INVENTORY',
  CREATE_PAYMENT = 'CREATE_PAYMENT',
  CREATE_ORDERS = 'CREATE_ORDERS',
  CONFIRM_VOUCHERS = 'CONFIRM_VOUCHERS',
  REMOVE_CART_ITEMS = 'REMOVE_CART_ITEMS',
  PROCESS_WALLET_PAYMENT = 'PROCESS_WALLET_PAYMENT',
}